+++ src/polymer/polymer-element-mixin.ts
@@ -14,12 +14,13 @@
 import * as dom5 from 'dom5';
 import * as estree from 'estree';
 
+import {Attribute, Event} from '../index';
 import {Annotation as JsDocAnnotation} from '../javascript/jsdoc';
 import {Document, ElementMixin, Method, Privacy, ScannedElementMixin, ScannedMethod, ScannedReference, SourceRange} from '../model/model';
 
 import {ScannedBehaviorAssignment} from './behavior';
 import {getOrInferPrivacy} from './js-utils';
-import {addMethod, addProperty, LocalId, Observer, PolymerExtension, PolymerProperty, ScannedPolymerExtension, ScannedPolymerProperty} from './polymer-element';
+import {_overwriteInherited, addMethod, addProperty, applyMixins, LocalId, Observer, PolymerExtension, PolymerProperty, ScannedPolymerExtension, ScannedPolymerProperty} from './polymer-element';
 
 export interface Options {
   name: string;
@@ -34,24 +35,38 @@ export interface Options {
 
 export class ScannedPolymerElementMixin extends ScannedElementMixin implements
     ScannedPolymerExtension {
-  properties: ScannedPolymerProperty[] = [];
-  methods: ScannedMethod[] = [];
-  observers: Observer[] = [];
-  listeners: {event: string, handler: string}[] = [];
-  behaviorAssignments: ScannedBehaviorAssignment[] = [];
+  readonly properties: ScannedPolymerProperty[] = [];
+  readonly methods: ScannedMethod[] = [];
+  readonly observers: Observer[] = [];
+  readonly listeners: {event: string, handler: string}[] = [];
+  readonly behaviorAssignments: ScannedBehaviorAssignment[] = [];
   // FIXME(rictic): domModule and scriptElement aren't known at a file local
   //     level. Remove them here, they should only exist on PolymerElement.
   domModule: dom5.Node|undefined = undefined;
   scriptElement: dom5.Node|undefined = undefined;
   pseudo: boolean = false;
-  abstract: boolean = false;
+  readonly abstract: boolean = false;
+  readonly sourceRange: SourceRange;
 
-  constructor(options?: Options) {
+  constructor({
+    name,
+    jsdoc,
+    description,
+    summary,
+    privacy,
+    sourceRange,
+    mixins,
+    astNode
+  }: Options) {
     super();
-    // TODO(justinfagnani): fix this constructor to not be crazy, or remove
-    // class altogether.
-    const optionsCopy = Object.assign({}, options) as Options;
-    Object.assign(this, optionsCopy);
+    this.name = name;
+    this.jsdoc = jsdoc;
+    this.description = description;
+    this.summary = summary;
+    this.privacy = privacy;
+    this.sourceRange = sourceRange;
+    this.mixins = mixins;
+    this.astNode = astNode;
   }
 
   addProperty(prop: ScannedPolymerProperty) {
@@ -59,25 +74,13 @@ export class ScannedPolymerElementMixin extends ScannedElementMixin implements
   }
 
   addMethod(method: ScannedMethod) {
+    // methods are only public by default if they're documented.
+    method.privacy = getOrInferPrivacy(method.name, method.jsdoc, true);
     addMethod(this, method);
   }
 
   resolve(document: Document): PolymerElementMixin {
-    const element = new PolymerElementMixin();
-    Object.assign(element, this);
-
-    for (const method of element.methods) {
-      // methods are only public by default if they're documented.
-      method.privacy = getOrInferPrivacy(method.name, method.jsdoc, true);
-    }
-    element.mixins = [];
-    for (const mixin of this.mixins) {
-      // TODO(rictic): we should mix these mixins into `this`. See
-      // PolymerElement's logic for this.
-      element.mixins.push(mixin.resolve(document));
-    }
-
-    return element;
+    return new PolymerElementMixin(this, document);
   }
 }
 
@@ -88,19 +91,79 @@ declare module '../model/queryable' {
 }
 export class PolymerElementMixin extends ElementMixin implements
     PolymerExtension {
-  properties: PolymerProperty[];
-  methods: Method[];
+  readonly properties: PolymerProperty[];
+  readonly methods: Method[];
 
-  observers: Observer[];
-  listeners: {event: string, handler: string}[];
-  behaviorAssignments: ScannedBehaviorAssignment[] = [];
-  domModule?: dom5.Node;
-  scriptElement?: dom5.Node;
-  localIds: LocalId[] = [];
+  readonly observers: Observer[];
+  readonly listeners: {event: string, handler: string}[];
+  readonly behaviorAssignments: ScannedBehaviorAssignment[] = [];
+  readonly domModule?: dom5.Node;
+  readonly scriptElement?: dom5.Node;
+  readonly localIds: LocalId[] = [];
 
-  kinds = new Set(['element-mixin', 'polymer-element-mixin']);
+  readonly kinds = new Set(['element-mixin', 'polymer-element-mixin']);
+  readonly pseudo: boolean;
+  readonly abstract: boolean;
 
-  abstract?: boolean;
+  constructor(scannedMixin: ScannedPolymerElementMixin, document: Document) {
+    super();
+    this.abstract = scannedMixin.abstract;
+    this.astNode = scannedMixin.astNode;
+    this.demos = scannedMixin.demos;
+    this.description = scannedMixin.description;
+    this.domModule = scannedMixin.domModule;
+    this.jsdoc = scannedMixin.jsdoc;
+    this.name = scannedMixin.name;
+    this.privacy = scannedMixin.privacy;
+    this.pseudo = scannedMixin.pseudo;
+    this.scriptElement = scannedMixin.scriptElement;
+    this.slots = scannedMixin.slots;
+    this.sourceRange = scannedMixin.sourceRange;
+    this.summary = scannedMixin.summary;
+    this.warnings = scannedMixin.warnings;
+
+    this.identifiers.add(this.name);
+
+    this.behaviorAssignments = Array.from(scannedMixin.behaviorAssignments);
+    this.observers = Array.from(scannedMixin.observers);
+    this.properties = [];
+    this.attributes = [];
+    this.methods = [];
+    this.events = [];
+
+    applyMixins(this, scannedMixin, document);
+
+    // TODO(rictic): do this in a more sane place, ideally somewhere in a parent
+    // constructor.
+    _overwriteInherited(
+        this.properties,
+        scannedMixin.properties as PolymerProperty[],
+        undefined,
+        this.warnings,
+        this.sourceRange,
+        true);
+    _overwriteInherited(
+        this.attributes,
+        scannedMixin.attributes as Attribute[],
+        undefined,
+        this.warnings,
+        this.sourceRange,
+        true);
+    _overwriteInherited(
+        this.events,
+        scannedMixin.events as Event[],
+        undefined,
+        this.warnings,
+        this.sourceRange,
+        true);
+    _overwriteInherited(
+        this.methods,
+        scannedMixin.methods as Method[],
+        undefined,
+        this.warnings,
+        this.sourceRange,
+        true);
+  }
 
   emitPropertyMetadata(property: PolymerProperty) {
     const polymerMetadata: any = {};