+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.0.0-alpha.38] - 2017-04-13
 
 * [minor breaking change] Revert of a change from alpha.37. The experimental `_fork` method on Analyzer returns an Analyzer again, not a Promise of an Analyzer.
+* Mix mixins into mixins. A PolymerElementMixin now has all of the members it inherits other mixins it mixes.
+* Improved our modeling of inheritance:
+  * overriding inherited members now works correctly
+  * overriding a private member produces a Warning
 
 ## [2.0.0-alpha.37] - 2017-04-12